Fix a memory leak associated with JavaObject destructor
C# object destructor is not guaranteed to be run on the main thread, and AndroidJNI methods become no-ops when not attached to the JNI from the current thread. This CL ensures that the thread is attached before calling the cleanup code.

Also bumps the version to 1.0, as this is now used in production (Input SDK).